requirements.txt
anyio==4.1.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.3.0
async-lru==2.0.4
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
Brotli==1.0.9
certifi==2024.2.2
charset-normalizer==2.0.4
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.2.0
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
docker-pycreds==0.4.0
entrypoints==0.4
exceptiongroup==1.2.0
executing==1.2.0
filelock==3.13.1
fonttools==4.25.0
fqdn==1.5.1
gitdb==4.0.7
GitPython==3.1.37
gmpy2==2.1.2
idna==3.4
importlib-metadata==7.0.1
ipykernel==6.29.2
ipython==8.15.0
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.3
joblib==1.2.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter_client==7.4.9
jupyter_core==5.3.1
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_server==2.10.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
munkres==1.1.4
nbclient==0.9.0
nbconvert==7.11.0
nest-asyncio==1.5.7
networkx==3.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.24.3
opencv-python==4.9.0.80
overrides==7.4.0
packaging==23.1
pandas==2.2.1
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pickleshare==0.7.5
Pillow==9.3.0
pip==24.0
platformdirs==3.10.0
ply==3.11
prometheus-client==0.19.0
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.16.1
pyparsing==3.0.9
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-ignite==0.4.13
pytz==2024.1
PyYAML==6.0.1
pyzmq==24.0.1
referencing==0.31.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.0
scikit-learn==1.3.0
scipy==1.11.4
Send2Trash==1.8.2
sentry-sdk==1.9.0
setproctitle==1.2.2
setuptools==69.1.0
sip==6.7.12
six==1.16.0
smmap==4.0.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.2
sympy==1.12
terminado==0.18.0
threadpoolctl==2.2.0
tinycss2==1.2.1
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tornado==6.3.3
tqdm==4.66.2
traitlets==5.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.1.0
wandb==0.16.3
wcwidth==0.2.6
webcolors==1.13
websocket-client==1.6.4
wheel==0.42.0
win-inet-pton==1.1.0
zipp==3.17.0
